added Nft pallet, moved house_bidding from callable, updated struct
